completed rewrite of calc. of directed loop probs.
The separate assignments and vertex_types txt files are no longer
needed. The epsilon supplied in the jobfile is now taken to be on top of
the amount which is needed to retain positivity of all probs (the latter
is internally referred to as delta). It defaults to 0.01 and should be
finite in order to avoid divisions by zero.